ADDED: Validations

+++ src/products/products.controller.ts
@@ -1,7 +1,9 @@
-import { Controller, Get, Post, Delete, Param, HttpCode, Body, ValidationPipe, UsePipes } from '@nestjs/common';
+import { Controller, Get, Post, Delete, Param, HttpCode, Body, ValidationPipe, UsePipes, BadRequestException } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDto } from './dto/create-product.dto';
 import { Product } from './interfaces/product.interface';
+import { Validator } from 'class-validator';
+const validator = new Validator();
 
 @Controller('products')
 export class ProductsController {
@@ -14,7 +16,8 @@ export class ProductsController {
 
     @Get(':id')
     findOne(@Param('id') id: string) {
-        return this.findOne(id);
+        if (!validator.isMongoId(id)) throw new BadRequestException();
+        return this.productsService.findOne(id);
     }
 
     @Post()
@@ -26,6 +29,7 @@ export class ProductsController {
     @Delete(':id')
     @HttpCode(204)
     deleteById(@Param('id') id: string) {
+        if (!validator.isMongoId(id)) throw new BadRequestException();
         this.productsService.deleteById(id);
     }
 }
